Move character coordinates utils to drafts
