+++ src/drafts/utils/character-coordinates.ts
